+++ topics/go/algorithms/bits/iseven/iseven.go
